m/dictionary/requirements.txt
aniso8601==3.0.2
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
isort==4.3.20
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
pep8==1.7.1
promise==2.2.1
psycopg2==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
virtualenv==16.6.0
wrapt==1.11.1
